Refactor infrastructure
