+++ src/compositionRoot.js
@@ -1,16 +1,12 @@
+import { ExportDatasetsUseCase } from "./domain/usecases/ExportDatasetsUseCase.js";
 import { DHISRepository } from "./data/repositories/DHISRepository.js";
-import { GetSelectedDataSetsUseCase } from "./domain/usecases/GetSelectedDataSetsUseCase.js";
-import { CreateXlsxFilesUseCase } from "./domain/usecases/CreateXlsxFilesUseCase.js";
+import { XLSXRepository } from "./data/repositories/XLSXRepository.js";
 
 export function getCompositionRoot() {
 	const dhisRepository = new DHISRepository();
+	const xlsxRepository = new XLSXRepository();
 
 	return {
-		dataSets: {
-			getSelected: new GetSelectedDataSetsUseCase(dhisRepository),
-		},
-		export: {
-			createFiles: new CreateXlsxFilesUseCase(),
-		},
+		exportToXlsx: new ExportDatasetsUseCase(dhisRepository, xlsxRepository),
 	};
 }
